+++ backend/canisters/community/impl/src/updates/c2c_uninstall_bot.rs
@@ -1,6 +1,6 @@
 use crate::guards::caller_is_local_user_index;
 use crate::model::events::CommunityEventInternal;
-use crate::{RuntimeState, activity_notifications::handle_activity_notification, execute_update};
+use crate::{RuntimeState, activity_notifications::handle_activity_notification, execute_update_even_if_frozen};
 use canister_api_macros::update;
 use canister_tracing_macros::trace;
 use constants::OPENCHAT_BOT_USER_ID;
@@ -10,7 +10,7 @@ use types::{BotRemoved, OCResult, c2c_uninstall_bot::*};
 #[update(guard = "caller_is_local_user_index", msgpack = true)]
 #[trace]
 fn c2c_uninstall_bot(args: Args) -> Response {
-    execute_update(|state| c2c_uninstall_bot_impl(args, state)).into()
+    execute_update_even_if_frozen(|state| c2c_uninstall_bot_impl(args, state)).into()
 }
 
 fn c2c_uninstall_bot_impl(args: Args, state: &mut RuntimeState) -> OCResult {