+++ backend/canisters/community/impl/src/updates/enable_invite_code.rs
@@ -88,8 +88,6 @@ struct PrepareResult {
 }
 
 fn prepare(state: &RuntimeState) -> OCResult<PrepareResult> {
-    state.data.verify_not_frozen()?;
-
     let member = state.get_calling_member(None, true)?;
     if member.role().can_invite_users(&state.data.permissions) {
         Ok(PrepareResult {