+++ backend/canisters/community/impl/src/updates/remove_member_from_channel.rs
@@ -16,8 +16,6 @@ pub(crate) fn remove_member_from_channel_impl(
     ext_caller: Option<Caller>,
     state: &mut RuntimeState,
 ) -> OCResult {
-    state.data.verify_not_frozen()?;
-
     let caller = state.verified_caller(ext_caller)?;
 
     let channel = state.data.channels.get_mut_or_err(&channel_id)?;