+++ backend/canisters/group/impl/src/updates/decline_invitation.rs
@@ -1,4 +1,4 @@
-use crate::{RuntimeState, execute_update};
+use crate::{RuntimeState, execute_update_even_if_frozen};
 use canister_api_macros::update;
 use canister_tracing_macros::trace;
 use group_canister::decline_invitation::*;
@@ -7,7 +7,7 @@ use oc_error_codes::OCErrorCode;
 #[update(msgpack = true)]
 #[trace]
 fn decline_invitation(_args: Args) -> Response {
-    execute_update(decline_invitation_impl)
+    execute_update_even_if_frozen(decline_invitation_impl)
 }
 
 fn decline_invitation_impl(state: &mut RuntimeState) -> Response {