+++ backend/canisters/group/impl/src/updates/unfollow_thread.rs
@@ -11,8 +11,6 @@ fn unfollow_thread(args: Args) -> Response {
 }
 
 fn unfollow_thread_impl(args: Args, state: &mut RuntimeState) -> OCResult {
-    state.data.verify_not_frozen()?;
-
     let user_id = state.get_caller_user_id()?;
     let now = state.env.now();
     state